Include basic.css from sphinx
This is necessary so that Sass can import basic.css into the built CSS
itself, without relying on an additional HTTP GET.

basic.css is from
https://github.com/sphinx-doc/sphinx/blob/master/sphinx/themes/basic/static/basic.css_t

It is BSD licensed.

I had to manually interpolate some of the Jinja templating in the
original basic.css_t, and simply used the theme's default values.