Set up colour palette based on www.lsst.org's blue
The primary colour is #409ac7, which is sampled from the blocks on
www.lsst.org. Then the red and yellow come from a triad system (they
have the same saturation and brightness on the colour wheel).

I used the overlay method to get the dark grey and white colours:
https://www.smashingmagazine.com/2016/04/web-developer-guide-color/#creating-a-cohesive-color-palette

Finally, I mapped these colours to have some semantic meanings and
began to apply these colours into the original CSS.